Moved IGravitationalField, NewtonianField, and PlummerField into gravity/forces
